ThreeSceneService to split the scene logic a bit.
