+++ softvis3d-frontend/src/components/scene/visualization/scene/Camera.ts
