+++ softvis3d-frontend/src/components/scene/visualization/scene/Setup.ts
